REFACTOR: DB init, connect 분리
