+++ database/db.py
@@ -1,0 +1,4 @@
+from mongoengine import *
+
+def initialize_db(app):
+    connect('moa')  # MongoDB Connector
